+++ CHANGES.md
@@ -23,6 +23,7 @@ Changes in 9.0
 - Hide Comment Codes tip message if Comments unchecked for Marking Period in InputFinalGrades.php
 - Add Get Student Labels Form JS (Disable unchecked fieldset) in StudentLabels.fnc.php & StudentLabels.php
 - Fix PHP8.1 deprecated use PostgreSQL $db_connection global variable in database.inc.php & Grades/includes/Dashboard.inc.php
+- Don't Delete Gender & Birthday Student Fields in Fields.fnc.php
 
 Changes in 8.9.1
 ----------------